feat: session guard delete
[#183]

+++ packages/backend/src/session/session.controller.ts
@@ -1,6 +1,5 @@
-import { Controller, Delete, Res, UseGuards } from '@nestjs/common';
+import { Controller, Delete, Res } from '@nestjs/common';
 import { SessionService } from './session.service';
-import { SessionGuard } from './session.guard';
 import { SessionId } from './session.decorator';
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { Response } from 'express';
@@ -11,7 +10,6 @@ export class SessionController {
   constructor(private readonly sessionService: SessionService) {}
 
   @Delete()
-  @UseGuards(SessionGuard)
   @ApiOperation({ summary: '세션을 삭제합니다.' })
   @ApiResponse({
     status: 200,
@@ -21,6 +19,11 @@ export class SessionController {
     @SessionId() sessionId: string,
     @Res() response: Response,
   ) {
+    if (!sessionId) {
+      response.end();
+      return;
+    }
+
     response.clearCookie('sessionId');
     this.sessionService.deleteSession(sessionId);
     response.end();
